requirements.txt
pynacl>=1.5
cbor2>=5.6
